[test_OMTypedParser] rename file / use om_parser_typed()
